Update Steam Group Mass Kicker Script.user.js
=== Big Update! ===
Added the ability to save the selected users to the kick list.
Saved users persist between pages.
If a user on the current page is in the saved list their checkmark is greyed out.
Opening the Kick Menu displays the saved and selected users in the list.
Hovering over users in the Kick Menu lets you preview their profile.
Clicking on users in the Kick Menu will remove them from the kick list. (Users won't be removed if you click cancel after removing users from the list, Clicking save will save the current users displayed in the list to the saved list. Clicking clear will delete the list entirely.)